Add .env file, and save confidential information in .env file.
